adds some bug fixes to the flags

+++ flags.h
@@ -1,7 +1,7 @@
 #include "loader.h"
 
 // Activates the selected flag
-void set_flag(uint8_t* flags, int flag);
+void set_flag(uint8_t* flags, uint8_t flag);
 // Deactivates the selected flag
 void unset_flag(uint8_t* flags, uint8_t flag);
